Fully XML-ify careers and display them.
Careers are now really XML, not just HTML. This allows for a lot of flexibility in displaying them (separation of display and data). This also lays the groundwork for including multiple careers per career xml file. Which is important, because InDesign won't do multiple XML files to one layout page.