Refactor setup
Setup before wasn't very focused. It had a bit of configuration that wasn't needed (the shared jinja_environment) and the book definition, which wasn't called book. This change gets rid of the uneeded shared configuration and changes the definition of the text to be called black_stars_rise, i.e. the actual name of this book.