Add print layout xml.
Okay, this is a little crazy. To allow for reusability, there's now a compile step for the InDesign XML files. The build script takes the book definition and spits out InDesign ready files for it. The main benefit of this is that a career or setup that appears in more than one setting can be single-sourced.